+++ changelog.md
@@ -8,6 +8,8 @@
   * Now play between 2 and 5 times, no matter the Amplifier
   * Cooldown is now reduced by 8% per level of Tired (instead of being divided by the level)
 * Reduced animal cap by 25%
+* Falling on climbables now reduces fall damage by 40%
+  * This can now be disabled in the config
 
 ## 4.7.1
 ### Fixes